+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Features:
   - Support showing error summaries when inline_errors is enabled (@rosswilson)
   - Name is now optional when creating static controls
   - Keep original form helper methods with _without_bootstrap suffix (#123, @a2ikm)
+  - Added support for icons (@datWav)
 
 ## 2.1.1 (2014-04-20)